quest: get rid of the broken quest-event system
Often the events are triggered even before the
server has updated the quest data. This leads to
questlog refreshs that happen when there is nothing
to update yet and won't trigger again once the data
from the server is refreshed. Also it's very easy
for other addons to interfere with those events.
(like addons that auto-complete quest dialogs)

The solution for now is, to create a function as
minimal as possible that checks for changes inside
the quest objectives manually.